backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
chromadb==0.5.5
pypdf==5.1.0
pymongo==4.10.1
pytz==2024.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==23.0.0
gevent==24.11.1
posthog==3.5.0
openai==1.58.1
whitenoise==6.7.0
